refactor(sts): split rustfs client operations modules
